muchos cambios

+++ OrderOut/Repositorys/category/ICategoryRepository.cs
@@ -7,7 +7,7 @@ namespace OrderOut.Repositorys
     public interface ICategoryRepository
     {
         Task<List<Category>> GetAllCategories();
-        Task<Category?> GetCategory(int categoryId);
+        Task<Category?> GetCategory(long categoryId);
         Task<bool> CreateCategory(Category category);
         Task<bool> UpdateCategory(Category category);
         Task<bool> DeleteCategory(long categoryId);
